Check if reverting the settings resolves failing tests
