+++ flowfile_core/tests/flowfile/test_node_step.py
@@ -110,25 +110,6 @@ def test_example_data_from_main_output():
     assert data == [{'name': 'John', 'city': 'New York'}], 'Data should be the same'
 
 
-def get_dependency_example():
-    graph = create_graph()
-    graph = add_manual_input(graph, data=[{'name': 'John', 'city': 'New York'},
-            {'name': 'Jane', 'city': 'Los Angeles'},
-            {'name': 'Edward', 'city': 'Chicago'},
-            {'name': 'Courtney', 'city': 'Chicago'}]
-)
-    node_promise = input_schema.NodePromise(flow_id=1, node_id=2, node_type='unique')
-    graph.add_node_promise(node_promise)
-
-    node_connection = input_schema.NodeConnection.create_from_simple_input(from_id=1, to_id=2)
-    add_connection(graph, node_connection)
-    input_file = input_schema.NodeUnique(flow_id=1, node_id=2,
-                                         unique_input=transform_schema.UniqueInput(columns=['city'])
-                                         )
-    graph.add_unique(input_file)
-    return graph
-
-
 def test_hash_change_data_with_old_data():
     graph = get_dependency_example()
     graph.run_graph()
